+++ lib/end_of_life/ruby_version.rb
@@ -7,26 +7,40 @@ module EndOfLife
 
     DB_PATH = File.join(__dir__, "../end_of_life.json")
 
-    EOL = File.read(DB_PATH)
-      .then { |json| JSON.parse(json, symbolize_names: true) }
-      .filter { |version| Date.parse(version[:eol]) <= Date.today }
-      .map { |version| Gem::Version.new(version[:latest]) }
-      .max
+    class << self
+      def from_file(file_name:, content:, parser: Parser)
+        parser.parse_file(file_name: file_name, content: content)
+      end
 
-    def self.from_file(file_name:, content:, parser: Parser)
-      parser.parse_file(file_name: file_name, content: content)
+      def eol_versions_at(date)
+        all_versions.filter { |version| version.eol_date <= date }
+      end
+
+      def latest_eol(at: Date.today)
+        eol_versions_at(at).max
+      end
+
+      private
+
+      def all_versions
+        @all_versions ||= File
+          .read(DB_PATH)
+          .then { |json| JSON.parse(json, symbolize_names: true) }
+          .map { |version| new(version[:latest], eol_date: Date.parse(version[:eol])) }
+      end
     end
 
-    attr :version
+    attr :version, :eol_date
 
-    def initialize(version_string)
+    def initialize(version_string, eol_date: nil)
       @version = Gem::Version.new(version_string)
+      @eol_date = eol_date
 
       freeze
     end
 
-    def eol?
-      @version <= EOL
+    def eol?(at: Date.today)
+      self <= RubyVersion.latest_eol(at: at)
     end
 
     def <=>(other)