+++ CHANGELOG.md
@@ -1,5 +1,55 @@
 ## [Unreleased]
 
+### Added
+
+- Add search methods to RubyVersion
+
+It's possible to query the all the eol versions or latest one at any given time.
+
+```ruby
+EndOfLife::RubyVersion.eol_versions_at(Date.today)
+# =>
+# [#<EndOfLife::RubyVersion:0x00007f9b1300d858
+#   @eol_date=#<Date: 2021-03-31 ((2459305j,0s,0n),+0s,2299161j)>,
+#   @version=Gem::Version.new("2.5.9")>,
+#  #<EndOfLife::RubyVersion:0x00007f9b1300cea8
+#   @eol_date=#<Date: 2020-03-31 ((2458940j,0s,0n),+0s,2299161j)>,
+#   @version=Gem::Version.new("2.4.10")>,
+#  #<EndOfLife::RubyVersion:0x00007f9b1300cb10
+#   @eol_date=#<Date: 2019-03-31 ((2458574j,0s,0n),+0s,2299161j)>,
+#   @version=Gem::Version.new("2.3.8")>,
+#  #<EndOfLife::RubyVersion:0x00007f9b1300c5e8
+#   @eol_date=#<Date: 2018-03-31 ((2458209j,0s,0n),+0s,2299161j)>,
+#   @version=Gem::Version.new("2.2.10")>,
+#  #<EndOfLife::RubyVersion:0x00007f9b1300c020
+#   @eol_date=#<Date: 2017-03-31 ((2457844j,0s,0n),+0s,2299161j)>,
+#   @version=Gem::Version.new("2.1.10")>,
+#  #<EndOfLife::RubyVersion:0x00007f9b112efbb8
+#   @eol_date=#<Date: 2016-02-24 ((2457443j,0s,0n),+0s,2299161j)>,
+#   @version=Gem::Version.new("2.0.0.pre.p648")>,
+#  #<EndOfLife::RubyVersion:0x00007f9b112ef028
+#   @eol_date=#<Date: 2015-02-23 ((2457077j,0s,0n),+0s,2299161j)>,
+#   @version=Gem::Version.new("1.9.3.pre.p551")>]
+
+EndOfLife::RubyVersion.latest_eol # returns today's latest EOL version
+# =>
+# #<EndOfLife::RubyVersion:0x00007f9b1300d858
+#  @eol_date=#<Date: 2021-03-31 ((2459305j,0s,0n),+0s,2299161j)>,
+#  @version=Gem::Version.new("2.5.9")>
+
+# returns the latest EOL version at a given date
+EndOfLife::RubyVersion.latest_eol(at: Date.parse("2024-03-31"))
+# =>
+# #<EndOfLife::RubyVersion:0x00007f9b1300e7d0
+#  @eol_date=#<Date: 2024-03-31 ((2460401j,0s,0n),+0s,2299161j)>,
+#  @version=Gem::Version.new("3.0.3")>
+```
+
+### Changed
+
+- `EndOfLife::RubyVersion::EOL` constant was removed in favor of `EndOfLife::RubyVersion.latest_eol` method.
+
+
 ----
 
 ## [0.2.0]